Cleanup implicit 64-bit to 32-bit conversions batch 1
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/593836
Commit-Queue: Herb Derby <herb@google.com>

+++ include/private/SkVx.h
@@ -617,7 +617,7 @@ SI auto map(std::index_sequence<I...>,
     // So, stifle CFI in this function.
     __attribute__((no_sanitize("cfi")))
 #endif
-    { return fn(args[i]...); };
+    { return fn(args[static_cast<int>(i)]...); };
 
     return { lane(I)... };
 }
